refactor(test): 테스트 함수 유틸로 변경 및 결과값 데이터 배열 객체로 관리
