+++ src/test/testUtils/expectedLogContains.js
@@ -1,0 +1,7 @@
+const expectLogContains = (received, expectedLogs) => {
+  expectedLogs.forEach((log) => {
+    expect(received).toContain(log);
+  });
+};
+
+export default expectLogContains;
