+++ src/test/testUtils/getLogSpy.js
@@ -1,0 +1,10 @@
+import { MissionUtils } from '@woowacourse/mission-utils';
+
+const getLogSpy = () => {
+  const logSpy = jest.spyOn(MissionUtils.Console, 'print');
+  logSpy.mockClear();
+
+  return logSpy;
+};
+
+export default getLogSpy;
