+++ src/test/testUtils/mockQuestions.js
@@ -1,0 +1,13 @@
+import { MissionUtils } from '@woowacourse/mission-utils';
+
+const mockQuestions = (inputs) => {
+  MissionUtils.Console.readLineAsync = jest.fn();
+
+  MissionUtils.Console.readLineAsync.mockImplementation(() => {
+    const input = inputs.shift();
+
+    return Promise.resolve(input);
+  });
+};
+
+export default mockQuestions;
